Fixed #425 BUG: in normal mode, keystrokes are leaking to the site
Mode.finish must be done synchronously in Mode.handleMapKey, otherwise it will
fail to find Trie node for a repeat keydown event in some cases under Linux,
since map_node may have not been reset in time. And the reason to run mapping
code in setTimeout is not that strong so far.